--- src/main/java/com/dangol/dangolsonnimbackend/boss/dto/AuthCodeResponseDTO.java
@@ -1,12 +1,0 @@
-package com.dangol.dangolsonnimbackend.boss.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthCodeResponseDTO {
-    private String authCode;
-}
