+++ src/main/java/com/dangol/dangolsonnimbackend/boss/dto/reponse/BossSigninResponseDTO.java
@@ -1,4 +1,4 @@
-package com.dangol.dangolsonnimbackend.boss.dto;
+package com.dangol.dangolsonnimbackend.boss.dto.reponse;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
