+++ src/test/java/com/dangol/dangolsonnimbackend/boss/repository/BossRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.dangol.dangolsonnimbackend.boss.repository;
 
 import com.dangol.dangolsonnimbackend.boss.domain.Boss;
-import com.dangol.dangolsonnimbackend.boss.dto.BossSignupRequestDTO;
+import com.dangol.dangolsonnimbackend.boss.dto.request.BossSignupRequestDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
