+++ src/test/java/com/dangol/dangolsonnimbackend/boss/repository/dsl/BossQueryRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.dangol.dangolsonnimbackend.boss.repository.dsl;
 
 import com.dangol.dangolsonnimbackend.boss.domain.Boss;
-import com.dangol.dangolsonnimbackend.boss.dto.BossSignupRequestDTO;
+import com.dangol.dangolsonnimbackend.boss.dto.request.BossSignupRequestDTO;
 import com.dangol.dangolsonnimbackend.boss.repository.BossRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
